pyproject.toml
[project]
name = "chatmodel-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
